+++ DrawCurve.Domen/Models/Core/Color.cs
@@ -10,9 +10,9 @@
             this.B = B;
             this.A = A;
         }
-        public byte R { get; set; }
-        public byte G { get; set; }
-        public byte B { get; set; }
-        public byte A { get; set; }
+        public int R { get; set; }
+        public int G { get; set; }
+        public int B { get; set; }
+        public int A { get; set; }
     }
 }
